[runtime] Fix waiting error reporting on WAIT_FAILURE return
We were checking for duplicates before we were checking if we exited early.
This means that in the case that we exit early enough, we sort an array of NULLs,
report that we have duplicates, and return with the wrong error message.

This does not change current behavior, as both exit paths return WAIT_FAILURE.